Replaced old laser scan with new one
