--- Assets/Scripts/LaserScanNew.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0efd25cf78f82221f954cc54c172d101
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
