+++ Assets/Scripts/LaserScanSensor.cs
@@ -7,13 +7,14 @@ using UnityEngine;
 using Unity.Robotics.ROSTCPConnector;
 using UnityEngine.Serialization;
 
+
 public class LaserScanSensor : MonoBehaviour
 {
     public string topic;
     [FormerlySerializedAs("TimeBetweenScansSeconds")]
     public double PublishPeriodSeconds = 0.1;
     public float RangeMetersMin = 0;
-    
+
     // burger
     // public float RangeMetersMax = 3.5f;
     // public float ScanAngleStartDegrees = 0;
@@ -28,11 +29,8 @@ public class LaserScanSensor : MonoBehaviour
 
     // Change the scan start and end by this amount after every publish
     public float ScanOffsetAfterPublish = 0f;
-    public float TimeBetweenMeasurementsSeconds = 0f;
+    
     public string frameId = "base_scan";
-
-    float m_CurrentScanAngleStart;
-    float m_CurrentScanAngleEnd;
     ROSConnection m_Ros;
     double m_TimeNextScanSeconds = -1;
     int m_NumMeasurementsTaken;
@@ -46,13 +44,10 @@ public class LaserScanSensor : MonoBehaviour
         m_Ros = ROSConnection.GetOrCreateInstance();
         m_Ros.RegisterPublisher<LaserScanMsg>(topic);
 
-        m_CurrentScanAngleStart = ScanAngleStartDegrees;
-        m_CurrentScanAngleEnd = ScanAngleEndDegrees;
-
         m_TimeNextScanSeconds = Clock.Now + PublishPeriodSeconds;
     }
 
-    void BeginScan()
+    private void BeginScan()
     {
         isScanning = true;
         m_TimeLastScanBeganSeconds = Clock.Now;
@@ -60,13 +55,13 @@ public class LaserScanSensor : MonoBehaviour
         m_NumMeasurementsTaken = 0;
     }
 
-    public void EndScan()
+    private void EndScan()
     {
         if (ranges.Count == 0)
         {
             Debug.LogWarning($"Took {m_NumMeasurementsTaken} measurements but found no valid ranges");
         }
-        else if (ranges.Count != m_NumMeasurementsTaken || ranges.Count != NumMeasurementsPerScan)
+        else if (ranges.Count != NumMeasurementsPerScan)
         {
             Debug.LogWarning($"Expected {NumMeasurementsPerScan} measurements. Actually took {m_NumMeasurementsTaken}" +
                              $"and recorded {ranges.Count} ranges.");
@@ -74,8 +69,10 @@ public class LaserScanSensor : MonoBehaviour
 
         TimeStamp timestamp = new TimeStamp(Clock.time);
         // Invert the angle ranges when going from Unity to ROS
-        var angleStartRos = -m_CurrentScanAngleStart * Mathf.Deg2Rad;
-        var angleEndRos = -m_CurrentScanAngleEnd * Mathf.Deg2Rad;
+        var angleStartRos = -ScanAngleStartDegrees * Mathf.Deg2Rad;
+        var angleEndRos = -ScanAngleEndDegrees * Mathf.Deg2Rad;
+        // var angleStartRos = -ScanAngleStartDegrees * Mathf.Deg2Rad;
+        // var angleEndRos = -ScanAngleEndDegrees * Mathf.Deg2Rad;
         if (angleStartRos > angleEndRos)
         {
             Debug.LogWarning("LaserScan was performed in a clockwise direction but ROS expects a counter-clockwise scan, flipping the ranges...");
@@ -101,7 +98,7 @@ public class LaserScanSensor : MonoBehaviour
             angle_min = angleStartRos,
             angle_max = angleEndRos,
             angle_increment = (angleEndRos - angleStartRos) / NumMeasurementsPerScan,
-            time_increment = TimeBetweenMeasurementsSeconds,
+            time_increment = 0f,
             scan_time = (float)PublishPeriodSeconds,
             intensities = new float[ranges.Count],
             ranges = ranges.ToArray(),
@@ -119,13 +116,34 @@ public class LaserScanSensor : MonoBehaviour
                              $"scheduled to start: {m_TimeNextScanSeconds:F}, rescheduling to now ({now:F})");
             m_TimeNextScanSeconds = now;
         }
+    }
 
-        m_CurrentScanAngleStart += ScanOffsetAfterPublish;
-        m_CurrentScanAngleEnd += ScanOffsetAfterPublish;
-        if (m_CurrentScanAngleStart > 360f || m_CurrentScanAngleEnd > 360f)
+    private void DoScan() 
+    {
+        float yawBaseDegrees = transform.rotation.eulerAngles.y;
+        float yawDegrees = ScanAngleStartDegrees + yawBaseDegrees + ScanOffsetAfterPublish;
+        float angleIncrement = (ScanAngleEndDegrees - ScanAngleStartDegrees) / NumMeasurementsPerScan;
+
+        for (int numMeasurements = 0; numMeasurements < NumMeasurementsPerScan; numMeasurements++)
         {
-            m_CurrentScanAngleStart -= 360f;
-            m_CurrentScanAngleEnd -= 360f;
+            var directionVector = Quaternion.Euler(0f, yawDegrees, 0f) * Vector3.forward;
+            var measurementStart = RangeMetersMin * directionVector + transform.position;
+            var measurementRay = new Ray(measurementStart, directionVector);
+            // Only record measurement if it's within the sensor's operating range
+            var foundValidMeasurement = Physics.Raycast(measurementRay, out var hit, RangeMetersMax);
+            // Even if Raycast didn't find a valid hit, we still count it as a measurement
+            if (foundValidMeasurement)
+            {
+                ranges.Add(hit.distance);
+            }
+            else
+            {
+                ranges.Add(float.MaxValue);
+            }
+
+            // Update scan degree
+            yawDegrees += angleIncrement;
+            ++m_NumMeasurementsTaken;
         }
     }
 
@@ -139,46 +157,8 @@ public class LaserScanSensor : MonoBehaviour
             }
 
             BeginScan();
-        }
-
-
-        var measurementsSoFar = TimeBetweenMeasurementsSeconds == 0 ? NumMeasurementsPerScan :
-            1 + Mathf.FloorToInt((float)(Clock.time - m_TimeLastScanBeganSeconds) / TimeBetweenMeasurementsSeconds);
-        if (measurementsSoFar > NumMeasurementsPerScan)
-            measurementsSoFar = NumMeasurementsPerScan;
-
-        var yawBaseDegrees = transform.rotation.eulerAngles.y;
-        while (m_NumMeasurementsTaken < measurementsSoFar)
-        {
-            var t = m_NumMeasurementsTaken / (float)NumMeasurementsPerScan;
-            var yawSensorDegrees = Mathf.Lerp(m_CurrentScanAngleStart, m_CurrentScanAngleEnd, t);
-            var yawDegrees = yawBaseDegrees + yawSensorDegrees;
-            var directionVector = Quaternion.Euler(0f, yawDegrees, 0f) * Vector3.forward;
-            var measurementStart = RangeMetersMin * directionVector + transform.position;
-            var measurementRay = new Ray(measurementStart, directionVector);
-            var foundValidMeasurement = Physics.Raycast(measurementRay, out var hit, RangeMetersMax);
-            // Only record measurement if it's within the sensor's operating range
-            if (foundValidMeasurement)
-            {
-                ranges.Add(hit.distance);
-            }
-            else
-            {
-                ranges.Add(float.MaxValue);
-            }
-
-            // Even if Raycast didn't find a valid hit, we still count it as a measurement
-            ++m_NumMeasurementsTaken;
-        }
-
-        if (m_NumMeasurementsTaken >= NumMeasurementsPerScan)
-        {
-            if (m_NumMeasurementsTaken > NumMeasurementsPerScan)
-            {
-                Debug.LogError($"LaserScan has {m_NumMeasurementsTaken} measurements but we expected {NumMeasurementsPerScan}");
-            }
+            DoScan();
             EndScan();
         }
-
     }
 }
